users_to_be_removed_modal is migrated
